- feat: 예외처리 로직 추가

+++ src/main/java/com/sprint/mission/discodeit/service/basic/BasicNotificationService.java
@@ -2,6 +2,8 @@ package com.sprint.mission.discodeit.service.basic;
 
 import com.sprint.mission.discodeit.dto.NotificationDTO;
 import com.sprint.mission.discodeit.entity.NotificationEntity;
+import com.sprint.mission.discodeit.exception.notification.ForbiddenNotificationAccessException;
+import com.sprint.mission.discodeit.exception.notification.NoSuchNotificationException;
 import com.sprint.mission.discodeit.mapper.NotificationMapper;
 import com.sprint.mission.discodeit.repository.NotificationRepository;
 import com.sprint.mission.discodeit.service.NotificationService;
@@ -10,6 +12,7 @@ import java.util.UUID;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 @Service
 @RequiredArgsConstructor
@@ -31,8 +34,16 @@ public class BasicNotificationService implements NotificationService {
 
   }
 
+  @Transactional
   @Override
   public void deleteNotificationById(UUID id, UUID userId) {
+
+    NotificationEntity notificationEntity = notificationRepository.findById(id)
+        .orElseThrow(NoSuchNotificationException::new);
+
+    if (!notificationEntity.getReceiverId().equals(userId)) {
+      throw new ForbiddenNotificationAccessException();
+    }
 
     notificationRepository.deleteById(id);
